feat(frontend): adjust exchange mapping ws message
