+++ frontend/app/src/plugins/rui/index.ts
@@ -103,6 +103,7 @@ import {
   RiGitMergeLine,
   RiGithubLine,
   RiGovernmentLine,
+  RiGuideLine,
   RiHandCoinLine,
   RiHandHeartLine,
   RiHeart2Line,
@@ -417,6 +418,7 @@ export function createRuiPlugin(defaults: Partial<RuiOptions['defaults']>): RuiP
         RiWalletLine,
         RiInformation2Line,
         RiInputField,
+        RiGuideLine,
       ],
     },
     defaults,
